more refactor; homeybuild removed; flow cards single definition only
