--- .homeybuild/drivers/tcp_driver/pair/configure_ip.html
@@ -1,77 +1,0 @@
-<style type="text/css">
-    span label {
-        width: 100px
-    }
-</style>
-<script type="text/javascript">
-    // Set window options
-    var width	= 400;
-    var height	= 400;
-    var left	= (screen.width/2)-(width/2);
-    var top		= (screen.height/2)-(height/2);
-    var url;
-
-    // Set translated items
-    $("label[for='name']").text(__('pair.configuration.name') + ':')
-    $("label[for='ip_address']").text(__('pair.configuration.ip_address') + ':')
-    $("label[for='tcp_port']").text(__('pair.configuration.tcp_port') + ':')
-
-    function configure()
-    {
-        $('.ip-scan-err-msg').text('');
-
-        var name = $("#name").val();
-        var ipAddress = $("#ip_address").val();
-        var tcpPort = $("#tcp_port").val();
-
-        if(isNaN(tcpPort) || tcpPort <= 0 || tcpPort > 65535) {
-            $('.ip-scan-err-msg').text(__('pair.configuration.tcp_port_invalid'));
-            return;
-        }
-
-
-        var device = {
-            data: { // this data object is saved to- and unique for the device. It is passed on the get and set functions as 1st argument
-                id: "PortScan" + new Date().getTime() // something unique, so your driver knows which physical device it is. A MAC address or Node ID, for example. This is required
-            },
-            name: name, // the name for this device (optional),
-            settings: {
-                host: ipAddress,
-                tcp_port: parseInt(tcpPort)
-            }
-        };
-
-        Homey.emit( 'configure_ip', device, function ( err, result ) {
-            if(err) {
-                $('.ip-scan-err-msg').text(err)
-            }
-            else {
-                //Successfully connected
-                Homey.addDevice(device, function( err, result ){
-                    if( err )
-                        return console.error(err);
-                });
-
-                //Done pairing
-                Homey.done();
-            }
-        } );
-    }
-</script>
-
-<p id="authentication_text" data-i18n="pair.configuration.explanation"></p>
-<p class="form-group">
-  <span class="input-wrap">    
-    <label for="name"></label>      
-    <input type="text" id="name"></input>
-    
-    <label for="ip_address"></label>    
-    <input type="text" id="ip_address"></input>
-        
-    <label for="tcp_port"></label>
-    <input type="text" id="tcp_port"></input>
-  </span>
-</p>
-<button class="button" id="authentication_button" onclick="configure();" data-i18n="pair.configuration.configure" />
-<br /><br />
-<p class="ip-scan-err-msg" style="color: #ff6300;"></p>
