--- .homeybuild/readme.txt
@@ -1,6 +1,0 @@
-Need to know when a device is switched on or off or when a phone comes home or leaves then this is for you.
-Just make sure the device has a fixed IP address and the app will check if it is available on the network.
-
-Note: Some devices will disconnect from the network when they are not being used which can cause false alarms.
-
-This app was orinally developed by Terry Hendrix.
