+++ .homeycompose/flow/conditions/device_is_offline.json
@@ -1,0 +1,16 @@
+{
+    "title": {
+        "en": "Device is !{{offline|online}}",
+        "nl": "Apparaat is !{{offline|online}}",
+        "de": "Gerät ist !{{offline|online}}",
+        "pl": "Urządzenie jest !{{offline|online}}"
+    },
+    "args": [
+        {
+            "name": "device",
+            "type": "device",
+            "filter": "driver_id=ip_driver|tcp_driver"
+        }
+    ],
+    "deprecated": true
+}
