--- drivers/ip_driver/driver.flow.compose.json
@@ -1,46 +1,0 @@
-{
-	"triggers": [
-		{
-			"id": "ip_device_change",
-			"title": {
-				"en": "Device changed state",
-				"nl": "Toestand van apparaat gewijzigd",
-				"de": "Gerätezustand geändert"
-			},
-			"args": [],
-			"tokens": [
-				{
-					"name": "value",
-					"title": {
-						"en": "Device Online State",
-						"nl": "Apparaat Online Status",
-						"de": "Online-Status Gerät"
-					},
-					"type": "boolean",
-					"example": true
-				}
-			]
-		}
-	],
-	"conditions": [
-		{
-			"id": "ip_device_is_online",
-			"title": {
-				"en": "Device is !{{online|offline}}",
-				"nl": "Apparaat is !{{online|offline}}",
-				"de": "Gerät ist !{{online|offline}}"
-			},
-			"args": []
-		},
-		{
-			"id": "ip_device_is_offline",
-			"title": {
-				"en": "Device is !{{offline|online}}",
-				"nl": "Apparaat is !{{offline|online}}",
-				"de": "Gerät ist !{{offline|online}}"
-			},
-			"args": [],
-			"deprecated": true
-		}
-	]
-}
