+++ drivers/tcp_driver/device.js
@@ -1,16 +1,11 @@
 /*jslint node: true */
 'use strict';
 
-const Homey = require( 'homey' );
-var net = require("net");
-const TcpIpDevice = require('../tcp_ip_device');
+const TcpIpDevice = require('../../lib/device');
 
-class tcpDevice extends TcpIpDevice
-{
-    async onInit()
-    {
+module.exports = class tcpDevice extends TcpIpDevice {
+    async onInit() {
         console.info("Booting TCP device ", this.getName());
         super.onInit();
     }
 }
-module.exports = tcpDevice;
