--- drivers/tcp_driver/driver.flow.compose.json
@@ -1,44 +1,0 @@
-{
-    "triggers": [
-        {
-            "id": "device_change",
-            "title": {
-                "en": "Device changed state",
-                "nl": "Toestand van apparaat gewijzigd",
-				"de": "Zustand des Geräts geändert"
-            },
-            "args": [],
-            "tokens": [
-                {
-                    "name": "value",
-                    "title": {
-                        "en": "Device Online State"
-                    },
-                    "type": "boolean",
-                    "example": true
-                }
-            ]
-        }
-    ],
-    "conditions": [
-        {
-            "id": "device_is_online",
-            "title": {
-                "en": "Device is !{{online|offline}}",
-                "nl": "Apparaat is !{{online|offline}}",
-                "de": "Gerät ist !{{online|offline}}"
-            },
-            "args": []
-        },
-        {
-            "id": "device_is_offline",
-            "title": {
-                "en": "Device is !{{offline|online}}",
-                "nl": "Apparaat is !{{offline|online}}",
-                "de": "Gerät ist !{{offline|online}}"
-            },
-            "args": [],
-            "deprecated": true
-        }
-    ]
-}
