+++ drivers/tcp_driver/driver.js
@@ -1,43 +1,11 @@
 "use strict";
 
 const Homey = require('homey');
-var net = require("net");
+const ConnectivityDriver = require('../../lib/driver');
 
-// the `init` method is called when your driver is loaded for the first time
-class tcpDriver extends Homey.Driver
-{
-    async onInit()
-    {
+module.exports = class tcpDriver extends ConnectivityDriver {
+    async onInit() {
         console.info("Booting TCP driver");
-        this.device_changed_state_trigger = this.homey.flow.getDeviceTriggerCard('device_state_changed');
+        super.onInit();
     }
-
-    device_came_online(device)
-    {
-        let tokens = {
-            value: true
-        };
-        this.device_changed_state_trigger
-            .trigger(device, tokens)
-            .catch(this.error);
-    }
-
-    device_went_offline(device)
-    {
-        let tokens = {
-            value: true
-        };
-        this.device_changed_state_trigger
-            .trigger(device, tokens)
-            .catch(this.error);
-    }
-
-    // the `pair` method is called when a user start pairing
-    async onPairListDevices()
-    {
-        this.homey.app.updateLog("Pairing started");
-
-    }
-
 }
-module.exports = tcpDriver;
