+++ lib/device.js
